separating files, fixes #12; working on single file generator
