+++ lib/constants.js
@@ -1,0 +1,9 @@
+'use strict';
+/* globals module */
+
+module.exports = {
+  TYPE_NODE: 'node',
+  TYPE_MIXIN: 'mixin',
+  TYPE_INCLUDE: 'include',
+  TYPE_NONE: 'standalone-comment'
+};
