+++ test/fixtures/mixins.jade
@@ -14,3 +14,10 @@ mixin mixin1
 
 mixin mixin2(arg1, arg2)
   div this is a mixin #{arg1} #{arg2}
+
+
+//- @jadedoc
+  name: mixin3
+
+mixin mixin3
+  +mixin1
