+++ test/tmp-tmp.jade
@@ -1,0 +1,28 @@
+//- @jadedoc
+  name: inclusion tag
+  description: this is an inclusion tag
+
+include 'node'//- @jadedoc
+  name: mixin1
+
+mixin mixin1
+  | this is a mixin
+
+
+//- @jadedoc
+  name: mixin2
+  description: this is jade mixin 2 documentation
+  arguments: 
+    arg1: foo
+    arg2: faa
+
+mixin mixin2(arg1, arg2)
+  div this is a mixin #{arg1} #{arg2}//- @jadedoc
+  name: some-node
+
+div.some-node
+  | this is some node
+
+div.some-other-node
+  | this is some other node//- @jadedoc
+  name: Only doc
